+++ .env.example
@@ -58,6 +58,8 @@ PETDEX_ADMIN_NOTIFY_EMAIL=
 # Admin
 PETDEX_ADMIN_USER_IDS=
 NEXT_PUBLIC_PETDEX_ADMIN_USER_IDS=
+PETDEX_COLLABORATOR_USER_IDS=
+NEXT_PUBLIC_PETDEX_COLLABORATOR_USER_IDS=
 
 # Optional app URLs
 PETDEX_ALL_PETS_PACK_URL=
