+++ .env.mock
@@ -24,3 +24,5 @@ OPENAI_API_KEY=sk-mock
 ELEVENLABS_API_KEY=mock
 PETDEX_ADMIN_USER_IDS=user_mock_contributor
 NEXT_PUBLIC_PETDEX_ADMIN_USER_IDS=user_mock_contributor
+PETDEX_COLLABORATOR_USER_IDS=user_mock_contributor
+NEXT_PUBLIC_PETDEX_COLLABORATOR_USER_IDS=user_mock_contributor
