+++ src/app/[locale]/collaborator/page.tsx
@@ -1,39 +1,38 @@
-import { getTranslations } from "next-intl/server";
+import { notFound } from "next/navigation";
+
+import { auth } from "@clerk/nextjs/server";
+
+import { isCollaborator } from "@/lib/admin";
 
 import {
   SubmissionReviewQueue,
   type SubmissionReviewQueueSearchParams,
 } from "@/components/submission-review-queue";
 
-export async function generateMetadata({
-  params,
-}: {
-  params: Promise<{ locale: string }>;
-}) {
-  const { locale } = await params;
-  const t = await getTranslations({ locale, namespace: "admin.metadata" });
-
-  return {
-    title: t("title"),
-    robots: { index: false, follow: false },
-  };
-}
+export const metadata = {
+  title: "Petdex Collaborator",
+  robots: { index: false, follow: false },
+};
 
 export const dynamic = "force-dynamic";
 
-export default async function AdminPage({
+export default async function CollaboratorPage({
   params,
   searchParams,
 }: {
   params: Promise<{ locale: string }>;
   searchParams: Promise<SubmissionReviewQueueSearchParams>;
 }) {
+  const { userId } = await auth();
+  if (!isCollaborator(userId)) notFound();
+
   const { locale } = await params;
   const resolvedSearchParams = await searchParams;
   return (
     <SubmissionReviewQueue
       locale={locale}
       searchParams={resolvedSearchParams}
+      actionScope="collaborator"
     />
   );
 }
