+++ src/components/auth-badge.tsx
@@ -18,7 +18,11 @@ import { ChatCircleDotsIcon } from "@phosphor-icons/react/dist/ssr";
 import { ExternalLink } from "lucide-react";
 import { useLocale, useTranslations } from "next-intl";
 
-import { canEditWeChatQrClientSafe, isAdminClientSafe } from "@/lib/admin";
+import {
+  canAccessCollaboratorAreaClientSafe,
+  isAdminClientSafe,
+  isCollaboratorClientSafe,
+} from "@/lib/admin";
 import { withLocale } from "@/lib/locale-routing";
 import { cn } from "@/lib/utils";
 
@@ -92,11 +96,10 @@ function UserDropdown({ compact = false }: { compact?: boolean }) {
   const { user, isLoaded } = useUser();
   const { signOut, openUserProfile } = useClerk();
   const showAdmin = isAdminClientSafe(user?.id);
-  // Collaborator entry surfaces tools the user can actually edit
-  // (currently /collaborator/wechat-qr for Henry + admins). We show it
-  // for admins too — they get both Admin (full dashboard) and the
-  // Collaborator shortcut so they don't have to navigate through admin.
-  const showCollaborator = canEditWeChatQrClientSafe(user?.id);
+  const showCollaborator = canAccessCollaboratorAreaClientSafe(user?.id);
+  const collaboratorHref = isCollaboratorClientSafe(user?.id)
+    ? "/collaborator"
+    : "/collaborator/wechat-qr";
   const unread = useHeaderState().state.feedback.count;
   const t = useTranslations("header");
   const locale = useLocale();
@@ -213,9 +216,7 @@ function UserDropdown({ compact = false }: { compact?: boolean }) {
             </DropdownMenuItem>
           ) : null}
           {showCollaborator ? (
-            <DropdownMenuItem
-              render={<Link href={href("/collaborator/wechat-qr")} />}
-            >
+            <DropdownMenuItem render={<Link href={href(collaboratorHref)} />}>
               <QrCodeIcon weight="duotone" className="size-4" />
               {t("collaborator")}
             </DropdownMenuItem>
