+++ src/lib/admin.ts
@@ -1,13 +1,16 @@
 // Server-side: PETDEX_ADMIN_USER_IDS (private env). Authoritative for any
 // gate that actually performs an admin action.
-export function getAdminUserIds(): Set<string> {
-  const raw = process.env.PETDEX_ADMIN_USER_IDS ?? "";
+function parseUserIds(raw: string | undefined): Set<string> {
   return new Set(
-    raw
+    (raw ?? "")
       .split(",")
       .map((id) => id.trim())
       .filter(Boolean),
   );
+}
+
+export function getAdminUserIds(): Set<string> {
+  return parseUserIds(process.env.PETDEX_ADMIN_USER_IDS);
 }
 
 export function isAdmin(userId: string | null | undefined): boolean {
@@ -20,13 +23,7 @@ export function isAdmin(userId: string | null | undefined): boolean {
 // whether to show admin links. Visibility-only — every server route that
 // mutates state still re-checks via isAdmin().
 export function getPublicAdminUserIds(): Set<string> {
-  const raw = process.env.NEXT_PUBLIC_PETDEX_ADMIN_USER_IDS ?? "";
-  return new Set(
-    raw
-      .split(",")
-      .map((id) => id.trim())
-      .filter(Boolean),
-  );
+  return parseUserIds(process.env.NEXT_PUBLIC_PETDEX_ADMIN_USER_IDS);
 }
 
 export function isAdminClientSafe(userId: string | null | undefined): boolean {
@@ -35,22 +32,82 @@ export function isAdminClientSafe(userId: string | null | undefined): boolean {
 }
 
 const HENRY_USER_ID = process.env.HENRY_USER_ID;
-// Client-safe mirror — used by UI components to show / hide the
-// 'Manage WeChat QR' menu entry. Same authoritative check on the server.
 const PUBLIC_HENRY_USER_ID = process.env.NEXT_PUBLIC_HENRY_USER_ID;
 
-export function canEditWeChatQr(userId: string | null | undefined): boolean {
+export function getCollaboratorUserIds(): Set<string> {
+  return parseUserIds(process.env.PETDEX_COLLABORATOR_USER_IDS);
+}
+
+export function isCollaborator(userId: string | null | undefined): boolean {
   if (!userId) return false;
   if (isAdmin(userId)) return true;
-  if (HENRY_USER_ID && userId === HENRY_USER_ID) return true;
-  return false;
+  return getCollaboratorUserIds().has(userId);
+}
+
+export function getPublicCollaboratorUserIds(): Set<string> {
+  return parseUserIds(process.env.NEXT_PUBLIC_PETDEX_COLLABORATOR_USER_IDS);
+}
+
+export function isHenry(userId: string | null | undefined): boolean {
+  if (!userId) return false;
+  return Boolean(HENRY_USER_ID && userId === HENRY_USER_ID);
+}
+
+export function isHenryClientSafe(userId: string | null | undefined): boolean {
+  if (!userId) return false;
+  return Boolean(PUBLIC_HENRY_USER_ID && userId === PUBLIC_HENRY_USER_ID);
+}
+
+export function canAccessCollaboratorArea(
+  userId: string | null | undefined,
+): boolean {
+  return isCollaborator(userId) || isHenry(userId);
+}
+
+export function canAccessCollaboratorAreaClientSafe(
+  userId: string | null | undefined,
+): boolean {
+  return isCollaboratorClientSafe(userId) || isHenryClientSafe(userId);
+}
+
+export function isCollaboratorClientSafe(
+  userId: string | null | undefined,
+): boolean {
+  if (!userId) return false;
+  if (isAdminClientSafe(userId)) return true;
+  return getPublicCollaboratorUserIds().has(userId);
+}
+
+export function canEditWeChatQr(userId: string | null | undefined): boolean {
+  return isCollaborator(userId) || isHenry(userId);
 }
 
 export function canEditWeChatQrClientSafe(
   userId: string | null | undefined,
 ): boolean {
-  if (!userId) return false;
-  if (isAdminClientSafe(userId)) return true;
-  if (PUBLIC_HENRY_USER_ID && userId === PUBLIC_HENRY_USER_ID) return true;
-  return false;
+  return isCollaboratorClientSafe(userId) || isHenryClientSafe(userId);
+}
+
+export function canReviewPetSubmissions(
+  userId: string | null | undefined,
+): boolean {
+  return isCollaborator(userId);
+}
+
+export function canReviewPetRequests(
+  userId: string | null | undefined,
+): boolean {
+  return isCollaborator(userId);
+}
+
+export function canReviewCollectionRequests(
+  userId: string | null | undefined,
+): boolean {
+  return isCollaborator(userId);
+}
+
+export function canReviewMetadataEdits(
+  userId: string | null | undefined,
+): boolean {
+  return isCollaborator(userId);
 }
